perf: Refactor to use isLogoVisible within state that it's needed of navlogo child
